requirements.txt
torch==1.13.1
transformers==4.25.1
pillow==9.3.0
matplotlib==3.6.0
numpy==1.23.4
mplcursors==0.5